Subsea Monitoring/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.9.2
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.9.24=hca03da5_0
contourpy=1.3.1=pypi_0
cycler=0.12.1=pypi_0
expat=2.6.3=h313beb8_0
fonttools=4.55.0=pypi_0
kiwisolver=1.4.7=pypi_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_1
matplotlib=3.9.2=pypi_0
ncurses=6.4=h313beb8_0
numpy=2.1.3=pypi_0
openssl=3.0.15=h80987f9_0
packaging=24.2=pypi_0
pillow=11.0.0=pypi_0
pip=24.2=py312hca03da5_0
pyparsing=3.2.0=pypi_0
pyserial=3.5=pypi_0
python=3.12.7=h99e199e_0
python-dateutil=2.9.0.post0=pypi_0
readline=8.2=h1a28f6b_0
setuptools=75.1.0=py312hca03da5_0
six=1.16.0=pypi_0
sqlite=3.45.3=h80987f9_0
tk=0.1.0=pypi_0
tzdata=2024b=h04d1e81_0
wheel=0.44.0=py312hca03da5_0
xz=5.4.6=h80987f9_1
zlib=1.2.13=h18a0788_1
